Rename modules to use lowercase
