+++ stograde/formatters/__init__.py
@@ -1,5 +1,4 @@
 from .base import format_collected_data
-from .gist import format_assignment_gist as gist
 from .html import format_assignment_html as html
 from .markdown import format_assignment_markdown as markdown
 from .truncate import truncate
