+++ stograde/process_assignment/analysis_result.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from stograde.process_assignment.Assignment_Status import AssignmentStatus
+from .assignment_status import AssignmentStatus
 
 
 @dataclass
