+++ stograde/specs/__init__.py
@@ -1,7 +1,7 @@
 from .cache import delete_cache
-from .File_Options import FileOptions
+from .file_options import FileOptions
 from .load import load_all_specs, load_some_specs
-from .Spec import Spec
-from .Spec_File import SpecFile
-from .Supporting_File import SupportingFile
+from .spec import Spec
+from .spec_file import SpecFile
+from .supporting_file import SupportingFile
 from .util import check_architecture, check_dependencies, get_filenames
