+++ stograde/student/analyze.py
@@ -4,10 +4,10 @@ from typing import List
 
 from stograde.common import chdir
 from stograde.common import find_unmerged_branches_in_cwd
-from stograde.process_assignment.Assignment_Status import AssignmentStatus
-from stograde.process_assignment.Assignment_Type import AssignmentType, get_assignment_type
+from stograde.process_assignment.assignment_status import AssignmentStatus
+from stograde.process_assignment.assignment_type import AssignmentType, get_assignment_type
 from stograde.specs import get_filenames, Spec
-from stograde.student.Student_Result import StudentResult
+from stograde.student.student_result import StudentResult
 
 
 def analyze(student: StudentResult, specs: List[Spec], check_for_branches: bool, ci: bool):
